requirements.txt
Django==5.2
djangorestframework>=3.16.0
django-cors-headers>=4.0
django-modeltranslation>=0.19.14
python-dotenv>=1.0
pillow>=11.2.1
drf-yasg>=1.21
gunicorn>=21.0
django-polymorphic==3.1.0
django-redis>=5.4.0
djangorestframework-simplejwt>=5.5.0
redis>=5.2.1
requests>=2.32.3
setuptools==78.1.0
six==1.17.0
aiogram==3.4.1